feat: add wildcard expr
